+++ R/dashbioNeedlePlot.R
@@ -1,8 +1,8 @@
 # AUTO GENERATED FILE - DO NOT EDIT
 
-dashbioNeedlePlot <- function(id=NULL, domainStyle=NULL, mutationData=NULL, needleStyle=NULL, rangeSlider=NULL, xlabel=NULL, ylabel=NULL) {
+dashbioNeedlePlot <- function(id=NULL, clickData=NULL, domainStyle=NULL, mutationData=NULL, needleStyle=NULL, rangeSlider=NULL, selectedData=NULL, xlabel=NULL, ylabel=NULL) {
     
-    props <- list(id=id, domainStyle=domainStyle, mutationData=mutationData, needleStyle=needleStyle, rangeSlider=rangeSlider, xlabel=xlabel, ylabel=ylabel)
+    props <- list(id=id, clickData=clickData, domainStyle=domainStyle, mutationData=mutationData, needleStyle=needleStyle, rangeSlider=rangeSlider, selectedData=selectedData, xlabel=xlabel, ylabel=ylabel)
     if (length(props) > 0) {
         props <- props[!vapply(props, is.null, logical(1))]
     }
@@ -10,7 +10,7 @@ dashbioNeedlePlot <- function(id=NULL, domainStyle=NULL, mutationData=NULL, need
         props = props,
         type = 'NeedlePlot',
         namespace = 'dash_bio',
-        propNames = c('id', 'domainStyle', 'mutationData', 'needleStyle', 'rangeSlider', 'xlabel', 'ylabel'),
+        propNames = c('id', 'clickData', 'domainStyle', 'mutationData', 'needleStyle', 'rangeSlider', 'selectedData', 'xlabel', 'ylabel'),
         package = 'dashBio'
         )